Return structured description from MT940 parser #15
The description (aka. Verwendungszweck aka. SVWZ) in SWIFT statements can, but does not need to be structured. This code detects if it is structured and, if so, it constructs an additional field in the returned array called 'description'. The existing 'description_1' and 'description_2' are kept for compatibility. The 'description' field contains an associative array mapping identifiers like "EREF" and "SVWZ" to the corresponding values. If the description is not structured, this field will still contain an array with only the key "SVWZ".

Open problems: Detection of structuredness and detection of new types rely on the presence of the + character at the fifth position. Is there any way this could go wrong (i.e. other legitimate messages that contain that character)?

"Types" could be renamed to "Sections" or sth like that.