Require trailing delimiter when parsing segments
This is for consistency of the interface. If the parser implicitly removes a trailing delimiter, then a caller that does not intend to pass the delimiter will have it accidentally trimmed in case the *content* of the segment happens to end with the delimiter character. This happens, for instance, with the data nested in HNVSD.