Add support for anonymous segments
Bank servers send segments in their responses that are not implemented in this library, and that do not need to be implemented either. In order for the parsing to succeed, unknown segments cannot be rejected. Instead of silently dropping them, this commit adds an AnonymousSegment class that holds their contents, which can be useful for debugging. 

Move the logic that detects the segment type from BaseSegment::parse() into the Parser where it belongs. Add a new AnonymousSegment class, which is returned from the Parser in case the parsed segment type is not explicitly implemented.