Introduce BaseGeschaeftsvorfallparameter and let HITANS use it
Every 6-letter segment sent by the bank ending in S (i.e. HIxyzS) is a segment that describes how the client can send the corresponding HKxyz segment. This base class captures the common fields and allows treating these parameters in the same way in future, e.g. a single implementation of the logic that checks the maximum number.