Remove redundant information from class/interface doc comments
This commit removes information from the phpDoc comments at the top of each file that has one, and possibly the entire comment block if it becomes empty. Specifically:
- The class name (e.g. "Class FinTs" or "Interface SegmentInterface") just repeats the class definition in PHP syntax below. It is at risk of getting outdated when the class is renamed, and adds no value.
- The @package annotation just repeats the namespace name. There are legitimate use cases for this annotation in case the package structure differs from the namespace structure, but this is not the case in this library. In that case, the official recommendation is not to use the annotation: https://docs.phpdoc.org/references/phpdoc/tags/package.html
- The @author annotation usually only names the original author of the file, so it is strictly less accurate than the Git history.

Besides adding unnecessary lines of code and becoming outdated easily during refactorings, a doc comment that only contains redundant information gives the false sense that the class is documented even when it really is not. Stripping the comment makes it more apparent that many of the classes in this library lack documentation. Missing documentation will be added in a future commit. This commit also does not remove the doc comment from some classes that are superseded by others and will be deleted anyway.