Remove BPD/UPD from processResponse()
Normally, the bank/user parameter data should only be needed to construct a request. In the rare cases where it is needed to interpret the response, the respective BaseAction implementation can wire the necessary pieces of information through its own state, as is done for GetStatementOfAccount::bankName in this commit.