Don't send HKTAN with TAN mode 999, remove TAN mode guessing hack
The hack was introduced because Consorsbank rejects TAN mode 999, saying it does not know that TAN mode. As it turns out, despite the BPD declaring that single-step TAN mode (999) is not supported, the bank does support it for weak dialog (as the specification requires), but it gets confused when a HKTAN segment is sent in this case (even though the specification requires that too). The other bank for which we have integration tests (DKB) also tolerates the missing HKTAN, so this commit makes that the new default behavior of the library. Firstly, this makes the hack obsolete that was used to guess the TAN mode, because 999 now works after all. Secondly, it might work better for Postbank (see #124 and #200) because the guessing does not work there (there are multiple available TAN modes) and Postbank also declares that 999 is not supported, but then using another TAN mode actually requires a TAN (i.e. strong authentication).