Split SanitizingCLILogger to separate its two concerns, enforce use of SanitizingLogger in FinTsNew
Not everyone wants to log to CLI, but everyone should want to sanitize the log output including on production servers. So make the SanitizingLogger part of the main library (not just the `Tests` package) and automatically wrap any logger with it. This also means that API users do not have to explicitly pass in their options/credentials anymore.