Populate HKTAN.segmentkennung with the business transaction's segment name
The field is supposed to contain the segment identifier of the segment for which the authentication is needed. The specification requires setting it upon login for special PIN/TAN use cases, requires setting it to HKIDN for regular logins, but does not specify anything for regular requests.

Now Consorsbank starts requiring it to be set to HKKAZ for such requests, and possibly similarly for others. This commit changes the behavior accordingly. HKKAZ also still works for DKB.