Add compatibility with banks that have not migrated to the new PSD2 world
The only known such bank so far is ING-DiBa, whose website states that they will keep the old HBCI running, but it's read-only for main accounts (fully working for special accounts and stock depots). This commit introduces a dummy TanMode called `NoPsd2TanMode` that users can supply in place of a real TAN mode in order to signal that the bank does not support any PSD2. In this case, all the TAN-related logic in the library is bypassed. In particular, no HKTAN segments are sent to the bank. Also, the anonymous dialog for BPD synchronization is suppressed, such that BPD synchronization takes place during the normal login. This is feasible because the BPD was only needed before the login to resolve the TanModes, which is not necessary when there's no real TanMode anyway.

The behavior for all other banks (i.e. when NoPsd2TanMode is used) remains unchanged.