Remove support for @Index annotations in field descriptors
It was originally added for two reasons, both of which are not applicable anymore: Firstly, it seemed useful to skip some unnecessary fields in the specification and to be able to continue at a later field by specifying its absolute index. But in practice, it is much easier to just copy over all fields from the specification than to fiddle with index magic and off-by-one errors. Secondly, it seemed necessary to declare the indices of fields behind a repeated field, e.g. a repeated field of length 100 occupies indices 0 through 99 and must thus be followed by index 100. But this use case is now supported implicitly as long as the repeated field specifies the @Max annotation, which is now mandatory.