Update README.md, remove COMPATIBILITY.md
README.md now pertains to the new library implementation. Instead of tracking bank compatibility in a long list (where a cross can easily be a false positive because banks evolve their systems over time, and a missing cross can easily be a false negative because not all tested banks are reported there), just declare compatibility with all banks, and we'll fix remaining issues.