Prevent accidental serialization of FinTs
While several other classes in this library (all sub-classes of BaseAction) are designed to be used with serialize() and unserialize(), FinTs works differently in that it has a persist() function, whose output does not include the options and sensitive credentials, so they passed to the factory function separately. This commit prevents users from calling serialize() and unserialize() on FinTs, which would cause confusion and errors at best, or work and be a security risk at worst.