Remove angle brackets from SanitizingLogger placeholder
Many programs interpreted <PRIVATE> as HTML, which led to the placeholder being disallowed entirely or being "rendered" as DOM elements so that they effectively disappeared when copy-pasting the rendered output. Since FinTS messages encode their own length in various places, it is important to keep the placeholders exactly as they are.