Detect temporary PIN lock (3931) and account lock (3938) warnings
Kreissparkasse send 3938 as a soft warning instead of a hard error, so it needs explicit detection to make the login process fail with the correct error message. (Otherwise, the DialogInitialization fails because the response contained no UPD, but that's only a secondary error caused by the locked account, so users really need to see the first error to know what's going on.)
Similarly, Postbank sends 3931 as a soft warning and even sends a TAN to the user right away, which would lead to an uninformative "Unexpected TAN request" error from the library that doesn't mention the PIN lock.
Also add a unit tests that are sensitive to this change.