Detect temporary account lock warning (3938)
Kreissparkasse send this as a soft warning instead of a hard error, so it needs explicit detection to make the login process fail with the correct error message. (Otherwise, the DialogInitialization fails because the response contained no UPD, but that's only a secondary error caused by the locked account, so users really need to see the first error to know what's going on.)
Also add a unit test that's sensitive to this change.