Call maybeThrowError() from execute() and submitTan()
This also affects the behavior of login(). It takes the burden off the caller to remember and call ensureSuccess() separately. Though they still need to check for needsTan(), as per the new example code and documentation.

Also, stop wrapping ServerExceptions inside UnexpectedResponseExceptions within executeWeakDialogInitialization(), as this was just making it harder to access the ServerException's helpers like indicatesBadLoginData().