Remove per-action error handling
Summary: This is a breaking API change for the new library implementation (which hasn't been released yet). isSuccess() is now isDone() and ensureSuccess() is now ensureDone(). Other status indicators on BaseAction are removed, as they are now unnecessary (an action is either done or not, execute() throws all the errors). See #282 for discussion.

In practice, most library users will only want to execute() one action at a time, the library currently only supports one action anyway, and many banks might only support this too. This makes the distinction between per-action errors (where one action could fail and another could succeed within the same request) and request-level errors (where there was an issue with the overall request and none of the actions were even attempted) less important. Previously (see #289), the execute() and submitTan() functions were changed to immediately throw all action-level errors too. This now makes it unnecessary to store errors in the BaseAction, which simplifies a lot of code in the library. It also makes the library easier to use because no generic \Exception isn't thrown anywhere anymore.

- Remove BaseAction:$error. Remove the distinction between "available" and "success", just call it "done".
- Remove @throws \Exception from ensureDone() and from all action getters, as the only possible exceptions that these can still throw are \RuntimeExceptions.
- Remove the FinTs::sendRequestForAction() and ServerException::extractErrorsForReference() helper functions that were needed to split the exceptions apart and assign them to the action. This also simplifies some catch blocks elsewhere in the FinTs class.