#304 Fix parsing of segments with required but empty DEGs where all subfields are optional
Not sure if a similar fix is necessary for the DEG parsing function, though probably not because DEG field delimiters need to be present even for empty subfields for the indices of subsequent fields to match. So probably the existing $allowEmpy logic in the DEG parsing function handles the same case there, and thus this fix is only necessary on the segment level.