Patch checkDecoupledSubmission() to tolerate absent HITAN(2) on success
Fixes #412.

The spec technically mandates that such a HITAN is sent on success, but in practice some banks don't do so, forcing us to adjust. Fortunately, the presence/absence of the 3956 respnose code is sufficient to tell whether the submission is already complete or not, so we rely (only) on that now.