Factor out a FinTs::processServerResponse() helper
Currently it has only one call site -- calling it from submitTan or checkDecoupledSubmission doesn't seem to work out of the box, because it falsely believes that a TAN request is needed (again). But maybe the TAN request recognition can be fixed in the future so that it can also be used for those. That said, it's not an issue as long as the TAN request is always the last thing that happens (strictly after VoP polling and confirmation).

This will be used by the future VoP functions in FinTs.