Make sure that all actions properly serialize their request-related fields
Some maneuvers in the FinTs protocol (like multi-TAN, which we don't currently support, or VOP, which we're adding support for) require re-submitting the same action later, using an identical request. If the application wants to do this in a separate PHP process (because the maneuver requires interactions with the user, during which the PHP process can't be kept alive), then during the serialization+deserialization roundtrip, we need to ensure that at least those fields survive which are required to construct the request.

While this doesn't happen for all action types in practice (e.g. a simple `GetBalance` request will never ask for VOP and at least today doesn't ask for a TAN either), this commit implements proper serialization in all actions for consistency (so that if someone forks an action in the future, they don't forget about this).